Extract testhelpers from loopback test suite
